Changin from ionic with react to ionic with angular
DUE to a major issue with importing stencil components in ionic/react I swiched to ionic/angular for completing the project. Ionic/React seams to have a problem with types which I couldn't find. For not losing more time I made the switch and now it is working as expected.